--- examples/example_text.txt
@@ -1,15 +1,0 @@
-The advent of artificial intelligence (AI) has revolutionized numerous industries, from healthcare to finance and beyond. One of the most significant advances 
-in AI has been the development of natural language processing (NLP) models, which are designed to enable machines to understand and generate human language. 
-These models are being applied in a wide range of applications, including chatbots, voice assistants, translation services, and content summarization.
-
-In the healthcare industry, NLP models are used to analyze vast amounts of medical data, helping doctors to make more accurate diagnoses and treatment decisions. 
-Similarly, in finance, AI algorithms are helping analysts predict market trends and make more informed investment choices. As AI continues to evolve, the possibilities 
-for its application seem endless.
-
-One of the most exciting developments in NLP has been the rise of large transformer-based models, such as OpenAI's GPT-3 and Google's BERT. These models have set new 
-benchmarks for language understanding and generation, and they are capable of performing a wide variety of language-related tasks with impressive accuracy. However, 
-there are still challenges to overcome, such as addressing biases in AI models and ensuring their ethical use.
-
-In the future, AI-powered NLP models are expected to become even more sophisticated, with applications ranging from autonomous vehicles that can understand human 
-speech to virtual assistants that can seamlessly assist with complex tasks. Despite the challenges, the future of AI and NLP holds immense promise, with the potential 
-to change how humans interact with machines in unprecedented ways.
