+++ requirements.txt
@@ -3,3 +3,5 @@ click==8.1.8
 torch==2.6.0
 python-dotenv==1.1.0
 huggingface_hub==0.29.3
+Flask==2.2.3
+Werkzeug==2.2.2
