+++ templates/index.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Text Summarizer</title>
+    <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}">
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            padding: 20px;
+        }
+        textarea {
+            width: 100%;
+            height: 150px;
+            margin-bottom: 10px;
+            padding: 10px;
+            font-size: 16px;
+        }
+        button {
+            padding: 10px 15px;
+            font-size: 16px;
+            cursor: pointer;
+        }
+        .summary {
+            margin-top: 20px;
+            padding: 10px;
+            background-color: #f0f0f0;
+            border-radius: 5px;
+        }
+    </style>
+</head>
+<body>
+    <h1>Text Summarizer</h1>
+    <form action="/summarize" method="POST">
+        <textarea name="input_text" placeholder="Введите текст для суммаризации..."></textarea>
+        <br>
+        <button type="submit">Суммаризировать</button>
+    </form>
+
+    {% if summary %}
+    <div class="summary">
+        <h2>Резюме:</h2>
+        <p>{{ summary }}</p>
+    </div>
+    {% endif %}
+</body>
+</html>
